+++ utils/isMerged.js
@@ -1,16 +1,7 @@
-const PULL_REQUEST_MERGED = {
-  github: {
-    locations: ['action'],
-    condition: 'closed'
-  },
-  gitlab: {
-    locations: ['object_attributes', 'action'],
-    condition: 'merge'
-  }
-}
+import { mergedOptions } from '../options/git'
 
 const isMerged = (git, body) => {
-  const locations = PULL_REQUEST_MERGED[git].locations
+  const locations = mergedOptions[git].locations
   let state = body
   for (const location of locations) {
     state = state[location]
@@ -21,7 +12,7 @@ const isMerged = (git, body) => {
     conditionAdd = body.pull_request.merged
   }
 
-  return state === PULL_REQUEST_MERGED[git].condition && conditionAdd
+  return state === mergedOptions[git].condition && conditionAdd
 }
 
 export default isMerged
