+++ utils/regex.js
@@ -1,0 +1,6 @@
+
+const regex = (string) => {
+  return new RegExp(string, 'i')
+}
+
+export default regex
